Timeline rewrite, it good now

+++ src/AXE/AXETypes.hpp
@@ -82,7 +82,7 @@ enum Keys_ {
 };
 
 struct Bookmark {
-	float position;
+	uint64_t position;
 	std::string name;
 };
